app/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
psycopg2-binary==2.9.3
prometheus_flask_exporter==0.18.1
